Add /logging_config URL
Printing the logging config to stderr every time the app starts up is
annoying. It's also not useful when exploring the logging configuration
when the app is hosted by an external ASGI server.

The URL is only active if logging_config is installed, which is a
development dependency, therefore it won't be available in production.