Switch to dnspython
dnspython3 appears to be a fork of dnspython 0.9.24 with some linting
and other mechanical code modernization performed on it. The fork still
links to the original dnspython git repository, and only a single 0.10.0
release has been made without a changelog update. It therefore appears
to be unmaintained.

In the meantime, the original dnspython module has had a couple of new
minor releases, so is still under maintenance.